feat(zsh): add mkfile function
This function takes in any number of filepaths and creates these files,
while also creating any missing parent directories on the way.

The goal of this to avoid having to do `mkdir -p ...; cd ...; touch ...`
as this is annoying, instead you can now just do `mkfile ...` which is a bit quicker.

May also add the option to cd to the directory as well (maybe with --cd flag).